Workaround for packstack expecting sample conf
Packstack, if configured to setup Tempest as part of an OpenStack
deployment, seems to expect the sample configuration file to be
present in the repository. This sample configuration was removed
from the upstream source somewhat recently and thus causes
Packstack to fail. This fix adds the sample back to the repo as a
workaround until the Packstack puppet manifests can be modified to
generate the sample configuration as part of their run.